+++ projects/evo-ui-kit/src/lib/components/evo-sidebar/evo-sidebar.component.ts
@@ -17,6 +17,12 @@ export enum EvoSidebarStates {
     VISIBLE = 'visible',
 }
 
+export enum EvoSidebarSizes {
+    LARGE = 'large'
+}
+
+const relativeFooterClass = 'relative-footer';
+
 @Component({
     selector: 'evo-sidebar',
     styleUrls: ['./evo-sidebar.component.scss'],
@@ -36,6 +42,7 @@ export class EvoSidebarComponent implements OnDestroy, OnInit {
     @Input() backButton: boolean;
     @Input() id: string;
     @Input() header: string;
+    @Input() size: EvoSidebarSizes;
     @Input() relativeFooter: boolean;
 
     @Output() back: EventEmitter<void> = new EventEmitter<void>();
@@ -74,6 +81,20 @@ export class EvoSidebarComponent implements OnDestroy, OnInit {
 
     get currentState(): string {
         return this.isVisible ? EvoSidebarStates.VISIBLE : EvoSidebarStates.HIDDEN;
+    }
+
+    get totalClasses(): string[] {
+        const classes: string[] = [];
+
+        if (this.size) {
+            classes.push(this.size);
+        }
+
+        if (this.relativeFooter) {
+            classes.push(relativeFooterClass);
+        }
+
+        return classes;
     }
 
     closeSidebar(source: EvoSidebarCloseTargets) {